% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiEmp.R
\name{BiEmp}
\alias{BiEmp}
\title{Function: Compute the bivariate empirical joint probability}
\usage{
BiEmp(X, Y)
}
\arguments{
\item{X}{The vector of a monthly hydro-climatic variable of n years.}

\item{Y}{The vector of a monthly hydro-climatic variable of n years.}
}
\value{
The empirical joint probability time scale
}
\description{
Function: Compute the bivariate empirical joint probability
}
\examples{
X=runif(120, min = 0, max = 100)
Y=runif(120, min = 0, max = 100)
fit<-BiEmp(X,Y) 
}
\references{
Hao, Z., et al. (2017). "An integrated package for drought monitoring, prediction and analysis to aid drought modeling and assessment." Environmental Modelling & Software 91: 199-209.
}
