% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought-package.R
\docType{package}
\name{drought-package}
\alias{drought-package}
\title{drought: A package for drought monitoring, prediction and analysis.}
\description{
The drought package provides three categories of important functions:
Drought_functions, Threshold_functions and UMFreq
}
\details{
\tabular{ll}{
Package: \tab drought\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2020-01-30\cr
License: \tab GPL-3\cr
}
}
\section{drought functions}{

The Drought functions ...

The Threshold functions ...

The UMFreq functions ...
}

\references{
Van Loon, A. F. (2015). "Hydrological drought explained." Wiley Interdisciplinary Reviews: Water 2(4): 359-392.

Hao, Z., et al. (2017). "An integrated package for drought monitoring, prediction and analysis to aid drought modeling and assessment." Environmental Modelling & Software 91: 199-209.

Shiau, J.-T. and R. Modarres (2009). "Copula‐based drought severity‐duration‐frequency analysis in Iran." Meteorological Applications: A journal of forecasting, practical applications, training techniques and modelling 16(4): 481-489.
}
\author{
Ze Jiang

Maintainer: Ze Jiang zejiangunsw@gamil.com
}
