% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drought_functions.R
\name{drought}
\alias{drought}
\title{Function: Compute drought characteristics}
\usage{
drought(X, start, end, theta)
}
\arguments{
\item{X}{A vector of time series data}

\item{start}{Start of time series}

\item{end}{End of time series}

\item{theta}{Threshold of a drought event}
}
\value{

}
\description{
Function: Compute drought characteristics
}
\examples{

}
\references{
Van Loon, A. F. (2015). "Hydrological drought explained." Wiley Interdisciplinary Reviews: Water 2(4): 359-392.
}
