% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPPred.R
\name{ESPPred}
\alias{ESPPred}
\title{Function: Drought prediction with ESP method}
\usage{
ESPPred(X, Y, L = 1, m = 7, ts = 6)
}
\arguments{
\item{X}{is the monthly variables.}

\item{Y}{is the monthly variables.}

\item{L}{is the lead time.}

\item{m}{is the monthly variables.}

\item{ts}{is the monthly variables.}
}
\value{
The prediction of univariate and multivariate drought index
}
\description{
Function: Drought prediction with ESP method
}
\examples{
X=runif(120, min = 0, max = 100)
Y=runif(120, min = 0, max = 100)
ESPPred(X,Y,L=1,m=7,ts=6)
}
