% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold_functions.R
\name{FixThres}
\alias{FixThres}
\title{Function: Compute fixed drought threshold}
\usage{
FixThres(data, percentile)
}
\arguments{
\item{data}{A vector of time series data}

\item{percentile}{The percentile of FDC (exceedance)}
}
\value{

}
\description{
Function: Compute fixed drought threshold
}
\examples{

}
\references{
Van Loon, A. F. (2015). "Hydrological drought explained." Wiley Interdisciplinary Reviews: Water 2(4): 359-392.
}
