% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Function: Get linear regression equations}
\usage{
lm_eqn(df)
}
\arguments{
\item{df}{drought characteristics data variable}
}
\value{
A string of equation, including intercept, coefficients and statistics.
}
\description{
Function: Get linear regression equations
}
\examples{

}
