% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pooling_functions.R
\name{Minor}
\alias{Minor}
\title{Function: Remove minor droughts}
\usage{
Minor(drought, MIN = 1)
}
\arguments{
\item{drought}{drought characteristics data variable}

\item{MIN}{minimum drought duration, default 1.}

\item{dur}{drought duration variable}
}
\value{
A subset of droughts with duration larger than x days or months
}
\description{
Function: Remove minor droughts
}
\examples{

}
