% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMFreq.R
\name{UMFreq}
\alias{UMFreq}
\title{Function: Univariate and multivariate return period (Gumbel copula)}
\usage{
UMFreq(X, Y, EL = 1)
}
\arguments{
\item{X}{the drought properties or indices}

\item{Y}{the drought properties or indices}

\item{EL}{the average reocurrence time, default EL=1 (annual maxima)}
}
\value{
The univariate and multivariate return period
}
\description{
Function: Univariate and multivariate return period (Gumbel copula)
}
\examples{

}
\references{
Hao, Z., et al. (2017). "An integrated package for drought monitoring, prediction and analysis to aid drought modeling and assessment." Environmental Modelling & Software 91: 199-209.

Shiau, J.-T. and R. Modarres (2009). "Copula‐based drought severity‐duration‐frequency analysis in Iran." Meteorological Applications: A journal of forecasting, practical applications, training techniques and modelling 16(4): 481-489.
}
