% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniEmp.R
\name{UniEmp}
\alias{UniEmp}
\title{Function: Compute the univariate empirical joint probability (EMP)}
\usage{
UniEmp(X, dist = c("Weibull", "Gringorten"))
}
\arguments{
\item{X}{The vector of a monthly hydro-climatic variable of n years.}

\item{dist}{The function for the plotting position formula (Gringorten or Weibull).}
}
\value{
The univariate EMP
}
\description{
Function: Compute the univariate empirical joint probability (EMP)
}
\examples{
X=runif(100, min = 0, max = 100)
P<-UniEmp(X,dist = "Gringorten") 
barplot(t(P))
plot(1/P,X)
}
