% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold_functions.R
\name{VarThres}
\alias{VarThres}
\title{Function: Compute varying drought threshold}
\usage{
VarThres(data, percentile, dates, by = "month")
}
\arguments{
\item{data}{A vector of time series data}

\item{percentile}{The percentile of FDC (exceedance), range from (0,1)}

\item{dates}{A vector of date variable}

\item{by}{Time step of the input data set}
}
\value{

}
\description{
Function: Compute varying drought threshold
}
\examples{

}
\references{
Van Loon, A. F. (2015). "Hydrological drought explained." Wiley Interdisciplinary Reviews: Water 2(4): 359-392.
}
