% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pooling_functions.R
\name{Minor}
\alias{Minor}
\title{Function: Remove minor droughts}
\usage{
Minor(file, dur, MIN)
}
\arguments{
\item{file}{drought characteristics data variable}

\item{dur}{drought duration variable}

\item{MIN}{minimum drought duration}
}
\value{
A subset of droughts with duration larger than x days or months
}
\description{
Function: Remove minor droughts
}
\examples{

}
