% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pooling_functions.R
\name{Pooling}
\alias{Pooling}
\title{Function: Pooling droughts}
\usage{
Pooling(drought, IT = 1, by = "month")
}
\arguments{
\item{drought}{drought characteristics data variable}

\item{IT}{interevent time period}

\item{by}{time step}
}
\value{
Pooled droughts by combining drought events within interevent time period
}
\description{
Function: Pooling droughts
}
\examples{

}
\references{
Van Loon, A. F. (2015). "Hydrological drought explained." Wiley Interdisciplinary Reviews: Water 2(4): 359-392.
}
